requirements.txt
# === OKX Futures+Spot Auto-Trader Dependencies ===

# 거래소 통합 라이브러리 (OKX 현물+선물)
ccxt>=4.0.0

# 데이터 처리
pandas>=2.1.0
numpy>=1.24.0,<2

# 디스코드 알림
discord-webhook>=1.3.0

# 환경 변수 / 설정
python-dotenv>=1.0.0
pyyaml>=6.0

# 스케줄링
apscheduler>=3.10.0

# 로깅
loguru>=0.7.0

# 데이터베이스
sqlalchemy>=2.0.0

# 비동기 HTTP
aiohttp>=3.9.0

# 차트 시각화
mplfinance>=0.12.10
matplotlib>=3.7.0

# 테스트
pytest>=7.4.0
pytest-asyncio>=0.21.0
